+++ src/NewPicker/PickerPanel/index.tsx
@@ -1,10 +1,6 @@
 import classNames from 'classnames';
 import { useEvent, useMergedState, warning } from 'rc-util';
 import * as React from 'react';
-import { isSame } from '../../utils/dateUtil';
-import useLocale from '../hooks/useLocale';
-import { getTimeConfig } from '../hooks/useTimeConfig';
-import useToggleDates from '../hooks/useToggleDates';
 import type {
   CellRender,
   Components,
@@ -16,9 +12,13 @@ import type {
   SharedPanelProps,
   SharedTimeProps,
 } from '../../interface';
+import { isSame } from '../../utils/dateUtil';
+import { toArray } from '../../utils/miscUtil';
+import useLocale from '../hooks/useLocale';
+import { getTimeConfig } from '../hooks/useTimeConfig';
+import useToggleDates from '../hooks/useToggleDates';
 import PickerContext from '../PickerInput/context';
 import useCellRender from '../PickerInput/hooks/useCellRender';
-import { toArray } from '../util';
 import DatePanel from './DatePanel';
 import DateTimePanel from './DateTimePanel';
 import DecadePanel from './DecadePanel';
