+++ src/NewPicker/util.ts
@@ -1,11 +1,3 @@
-export function toArray<T>(val: T | T[]): T[] {
-  if (val === null || val === undefined) {
-    return [];
-  }
-
-  return Array.isArray(val) ? val : [val];
-}
-
 export function fillIndex<T extends any[]>(ori: T, index: number, value: T[number]): T {
   const clone = [...ori] as T;
   clone[index] = value;
