--- src/utils/getRanges.tsx
@@ -1,62 +1,0 @@
-import * as React from 'react';
-import type { Components, Locale, RangeList } from '../interface';
-
-export type RangesProps = {
-  prefixCls: string;
-  rangeList?: RangeList;
-  components?: Components;
-  needConfirmButton: boolean;
-  onNow?: null | (() => void) | false;
-  onOk?: null | (() => void) | false;
-  okDisabled?: boolean;
-  showNow?: boolean;
-  locale: Locale;
-};
-
-export default function getRanges({
-  prefixCls,
-  // rangeList = [],
-  components = {},
-  needConfirmButton,
-  onNow,
-  onOk,
-  okDisabled,
-  showNow,
-  locale,
-}: RangesProps) {
-  let presetNode: React.ReactNode;
-  let okNode: React.ReactNode;
-
-  if (needConfirmButton) {
-    const Button = (components.button || 'button') as any;
-
-    if (onNow && showNow !== false) {
-      presetNode = (
-        <li className={`${prefixCls}-now`}>
-          <a className={`${prefixCls}-now-btn`} onClick={onNow}>
-            {locale.now}
-          </a>
-        </li>
-      );
-    }
-
-    okNode = needConfirmButton && (
-      <li className={`${prefixCls}-ok`}>
-        <Button disabled={okDisabled} onClick={onOk}>
-          {locale.ok}
-        </Button>
-      </li>
-    );
-  }
-
-  if (!presetNode && !okNode) {
-    return null;
-  }
-
-  return (
-    <ul className={`${prefixCls}-ranges`}>
-      {presetNode}
-      {okNode}
-    </ul>
-  );
-}
